More header sorting

+++ include/JSystem/JMessage/TParse.h
@@ -15,17 +15,21 @@ namespace JMessage {
 struct TParse : public JGadget::binary::TParse_header_block {
 	TParse(TResourceContainer*);
 
-	virtual ~TParse();                                                          // _08
-	virtual bool parseHeader_next(const void**, unsigned long*, unsigned long); // _0C
-	virtual bool parseBlock_next(const void**, unsigned long*, unsigned long);  // _10
+	virtual ~TParse();                                      // _08
+	virtual bool parseHeader_next(const void**, u32*, u32); // _0C
+	virtual bool parseBlock_next(const void**, u32*, u32);  // _10
+
+	// _00 = VTBL
 };
 
 struct TParse_color : public JGadget::binary::TParse_header_block {
 	TParse_color(TResourceContainer*);
 
-	virtual ~TParse_color();                                                    // _08
-	virtual bool parseHeader_next(const void**, unsigned long*, unsigned long); // _0C
-	virtual bool parseBlock_next(const void**, unsigned long*, unsigned long);  // _10
+	virtual ~TParse_color();                                // _08
+	virtual bool parseHeader_next(const void**, u32*, u32); // _0C
+	virtual bool parseBlock_next(const void**, u32*, u32);  // _10
+
+	// _00 = VTBL
 };
 } // namespace JMessage
