+++ include/JSystem/JMessage/TReference.h
@@ -7,9 +7,10 @@ namespace JMessage {
 struct TResourceContainer;
 
 struct TReference {
-	virtual ~TReference();                      // _08
-	virtual char* do_word(unsigned long) const; // _0C
+	virtual ~TReference();            // _08
+	virtual char* do_word(u32) const; // _0C
 
+	// _00 = VTBL
 	TResourceContainer* _04; // _04
 };
 } // namespace JMessage
